Implement new capability-based scheduling algorithm
With the old algorithm, workflow execution could stall if there
was a process chain that could not be executed because of a missing
agent, even if there were other agents that would have been able to
execute the remaining process chains. The new algorithm executes
process chains in the order they were added to the registry and always
fetches as many of them as possible from the registry if there are
enough agents available that can execute them. Process chains that
cannot be executed because they require capabilities none of the
agents can provide are skipped but resumed immediately as soon as
an agent with the matching capabilities joins the cluster.