Notify all instances about left remote agent
Prior to this fix, the REMOTE_AGENT_LEFT message was only sent to
local consumers and only if agents.remove() returned a non-null
value. This means that the message was only published on the one
instance that removed the agent from the list. This could lead to
a RemoteAgent on another Steep instance to not get notified about
a left agent and therefore to not abort the process chain.