Never consider an agent as idle if there is runnin process chain
This prevents the agent from being marked as idle if there was a clock jump
on the system and the busy timer was not reset. Without this fix, it could
happen that, with a busy timeout of 1m and a clock jump of more than 1m, the
agent was marked as idle even though there was still a process
chain being executed by the executeProcessChain() method.